+++ local.tf
@@ -123,6 +123,6 @@ locals {
     helm_modify       = 600
   }
 
-  network_plugin = var.experimental.windows_support ? "azure" : "kubenet"
-  azure_env      = startswith(var.location, "usgov") ? "usgovernment" : var.azure_env
+  cni       = var.experimental.windows_support ? "azure" : "kubenet"
+  azure_env = startswith(var.location, "usgov") ? "usgovernment" : var.azure_env
 }